remove warnings

+++ KlayGE/Plugins/Src/Render/D3D11/D3D11Texture2D.cpp
@@ -160,14 +160,14 @@ namespace KlayGE
 
 	uint32_t D3D11Texture2D::Width(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return widthes_[level];
 	}
 
 	uint32_t D3D11Texture2D::Height(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return heights_[level];
 	}
@@ -175,7 +175,7 @@ namespace KlayGE
 	uint32_t D3D11Texture2D::Depth(int level) const
 	{
 		UNREF_PARAM(level);
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return 1;
 	}
