+++ KlayGE/Plugins/Src/Render/D3D9/D3D9Texture3D.cpp
@@ -130,21 +130,21 @@ namespace KlayGE
 
 	uint32_t D3D9Texture3D::Width(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return widths_[level];
 	}
 
 	uint32_t D3D9Texture3D::Height(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return heights_[level];
 	}
 
 	uint32_t D3D9Texture3D::Depth(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return depths_[level];
 	}
