+++ KlayGE/Plugins/Src/Render/D3D9/D3D9TextureCube.cpp
@@ -128,7 +128,7 @@ namespace KlayGE
 
 	uint32_t D3D9TextureCube::Width(int level) const
 	{
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return widths_[level];
 	}
@@ -141,7 +141,7 @@ namespace KlayGE
 	uint32_t D3D9TextureCube::Depth(int level) const
 	{
 		UNREF_PARAM(level);
-		BOOST_ASSERT(level < numMipMaps_);
+		BOOST_ASSERT(static_cast<uint32_t>(level) < numMipMaps_);
 
 		return 1;
 	}
